10/01/18 02:00AM

+++ WorkInProgress/Assets/Scripts/Character/Enemies/EnemyHPBar.cs
@@ -7,7 +7,7 @@ public class EnemyHPBar : MonoBehaviour
     public Attackable attackable;
 
     private float width;
-    private int maxHP;
+    private float maxHP;
 
     private void Awake()
     {
@@ -24,7 +24,7 @@ public class EnemyHPBar : MonoBehaviour
 
     private void UpdateHealthBar()
     {
-        int currentHP = attackable.GetHealth();
+        float currentHP = attackable.GetHealth();
 
         float newWidth = (width / maxHP) * currentHP;
